+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to `Laragram` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.5.3] (2022-11-28)
+
+### Added
+
+- Added the ability to set array patterns in `contains` route
+
+### Changed
+
+- Extend model `User` for Laravel Authentication and Factories
+- Fix bugs
+
 ## [v1.5.2] (2022-11-27)
 
 ### Added